backend/requirements.txt
# FastAPI framework and related
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.0
starlette==0.27.0
python-multipart>=0.0.6
sse-starlette>=1.6.5  # For server-sent events in FastAPI
email-validator>=2.0.0  # Required for pydantic EmailStr

# Monitoring and metrics
prometheus_client>=0.17.0  # For metrics collection

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
cryptography==41.0.5
PyJWT>=2.8.0  # JWT implementation for Python

# Utilities
python-dotenv>=1.0.0
pyyaml==6.0.1
httpx>=0.25.0
requests==2.31.0
jinja2==3.1.2
werkzeug>=2.3.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9
postgrest>=0.13.0  # For Supabase integration

# File handling
python-docx>=1.0.0
pdf2image==1.16.3
reportlab>=4.1.0,<5.0.0
fpdf2==2.7.6

# Data processing
numpy>=2.2.0  # Updated for Python 3.13 compatibility
pandas>=2.2.0  # Updated for Python 3.13 compatibility
pdfplumber>=0.10.3  # Updated to resolve version conflict
pytesseract==0.3.10

# Concurrency and performance
aiofiles==23.2.1
asyncio==3.4.3
psutil==5.9.6
aiohttp>=3.9.0
aiodns>=3.0.0  # For faster DNS resolution with aiohttp

# Date/time handling
pytz==2023.3.post1
python-dateutil==2.8.2

# Error tracking and logging
sentry-sdk==1.32.0
python-json-logger==2.0.7

# ---------------- Document / PDF Processing ----------------
docxtpl==0.16.7
pdfplumber>=0.10.3
Pillow>=10.0.1,<11.0.0
mammoth>=1.6.0
docx2pdf>=0.1.8  # Note: docx2pdf requires 'pythoncom' module (Windows-only)
docx2txt>=0.8
reportlab>=4.1.0,<5.0.0
PyMuPDF>=1.22.5,<2.0.0
pdfrw>=0.4
weasyprint>=60.2
python-magic>=0.4.27
# python-magic-bin is only needed on Windows
# python-magic-bin>=0.4.14;sys_platform=='win32'

# Note: For Linux deployment (like Render), ensure LibreOffice or unoconv is installed
# at the system level for PDF conversion: 
# apt-get update && apt-get install -y libreoffice

# ---------------- API / Networking ----------------
aiohttp>=3.9.0
PyPDF2>=3.0.0
sse-starlette>=1.6.5  # For server-sent events in FastAPI

# ---------------- Development Tools ----------------
black>=24.2.0
flake8>=7.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Make sure we have compatible versions for common dependencies
certifi>=2023.7.22
charset-normalizer>=3.0.0
idna>=3.4
urllib3>=1.26.0,<2.0.0  # Many packages expect urllib3 < 2.0.0
